#1 rate products feature [admin can't rate, user can only rate the products that they doesn't own] #2 return the categories in search api #3 delete order_product table

+++ app/Services/ProductService.php
@@ -6,6 +6,7 @@ use App\Enums\Role;
 use App\Facades\FileManager;
 use App\Http\Requests\ProductCreateRequest;
 use App\Http\Requests\ProductUpdateRequest;
+use App\Http\Requests\RateRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\Category;
 use App\Models\Product;
@@ -13,6 +14,7 @@ use App\Models\Store;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
@@ -144,13 +146,9 @@ class ProductService
         return $store->user->id == $user->id;
     }
 
-    private function ownProductOrAdmin(Store $store, User $user): void
+    private function ownProductOrAdmin(Store $store, User $user): bool
     {
-        throw_if(
-            !$this->ownership($store, $user) && !$user->hasRole(Role::ADMIN),
-            AccessDeniedHttpException::class,
-            'access denied',
-        );
+        return ($this->ownership($store, $user) || $user->hasRole(Role::ADMIN));
     }
 
     public function index()
@@ -250,7 +248,11 @@ class ProductService
 
         $store = $product->store;
 
-        $this->ownProductOrAdmin($store, $user);
+        throw_unless(
+            $this->ownProductOrAdmin($store, $user),
+            AccessDeniedHttpException::class,
+            'access denied',
+        );
 
         return $product->delete();
     }
@@ -281,6 +283,41 @@ class ProductService
         }
 
         return $query->get();
+    }
+
+    public function rate(Product $product, RateRequest $request)
+    {
+        $validated = $request->validated();
+
+        $user = User::find(Auth::id());
+
+        throw_if(
+            $this->ownProductOrAdmin($product->store, $user) || !$user->hasOrderedProduct($product),
+            AccessDeniedHttpException::class,
+            'access denied',
+        );
+
+        DB::transaction(function () use ($validated, $user, $product) {
+            if ($user->hasRatedProduct($product)) {
+                $rate = $user->rate($product);
+
+                $product->rate_sum += $validated['rate'] - $rate;
+
+                $user->rates()->updateExistingPivot($product->id, $validated);
+            } else {
+                $user->rates()->attach(
+                    $product->id,
+                    ['rate' => $validated['rate']]
+                );
+
+                $product->rate_sum += $validated['rate'];
+                $product->rate_count++;
+            }
+
+            $product->save();
+        });
+
+        return $product;
     }
 }
